ecommerce/ai-agent/requirements.txt
langchain
langgraph
langchain-google-genai
chainlit
sqlalchemy
pymysql
python-dotenv
